+++ src/main/kotlin/com/kamesuta/paintcraft/canvas/paint/PaintRect.kt
@@ -98,7 +98,7 @@ class PaintRect(override val session: CanvasSession) : PaintTool {
                             uvEnd.y,
                             color,
                             false,
-                            session.mode.thickness.toInt(),
+                            session.mode.thickness,
                         )
                     )
                 }
